requirements.txt
robotframework==4.1.1
robotframework-jsonlibrary==0.3.1
robotframework-requests==0.9.2
